services/upload-api/requirements.txt
fastapi==0.111.0
firebase-admin==6.5.0
psycopg2-binary==2.9.9
uvicorn[standard]==0.30.1
google-cloud-storage==2.18.2
google-cloud-pubsub==2.21.5
pydantic==2.7.3
pytest==8.3.2
